\name{ywGetToken}
\alias{ywGetToken}
\title{Get token for Yandex Webmaster API access}
\description{ywGetToken gets token for Yandex Webmaster API v 3.0 access. Allow access for application in browser, copy token and paste to RStudio.}
\usage{
ywGetToken()
}

\arguments{
  
}
\details{}
\value{Returns token for access to API.
}
\references{}
\author{Ivan Kutas}
\note{}

\seealso{}
\examples{
token <- yadirGetToken()
}

\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
