\name{ywGetUserId}
\alias{ywGetUserId}
\title{Get user id from Yandex Webmaster}
\description{ywGetUserId downloads user id from Yandex Webmaster API v 3.0}
\usage{
ywGetUserId(token = NULL)
}

\arguments{
  \item{token}{Token for Yandex Webmaster API from function ywGetToken}
}
\details{}
\value{Returns user id.}
\references{}
\author{Ivan Kutas}
\note{}

\seealso{}
\examples{
user_id <- ywGetUserId(token=myToken)
}

\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
